+++ src/components/request/Entry.tsx
@@ -1,16 +1,17 @@
 import React, { Dispatch, SetStateAction } from "react";
+import { useNavigate } from "react-router-dom";
 import { useRecoilState } from "recoil";
 import { requestBasicAtom } from "../../state/atoms";
 import Button from "../../shared/components/Button";
 import BasicInput from "./BasicInput";
 import PosterUploader from "./PosterUploader";
-import { StyledEntry } from "./styles/requestStyle";
+import { StyledEntry, StyledRequestModal } from "./styles/requestStyle";
 import PlaceInput from "./PlaceInput";
 import ArtistInputContainer from "./ArtistInputContainer";
 import GoodsInputContainer from "./GoodsInputContainer";
 import HashTagsContainer from "./HashTagsContainer";
-import Modal from "./Modal";
 import DateRangeInput from "./DateRangeInput";
+import Modal from "../../shared/components/Modal";
 
 type EntryProps = {
 	isModalOpen: boolean;
@@ -21,6 +22,8 @@ type EntryProps = {
 };
 
 const Entry = ({ isModalOpen, setModalOpen, setBottomSheetOpen, handleSubmit, resetAllStates }: EntryProps) => {
+	const navigate = useNavigate();
+
 	const [basicInputs, setBasicInputs] = useRecoilState(requestBasicAtom);
 	const { organizer, snsId, link } = basicInputs;
 
@@ -36,6 +39,18 @@ const Entry = ({ isModalOpen, setModalOpen, setBottomSheetOpen, handleSubmit, re
 			...prev,
 			[id]: "",
 		}));
+	};
+
+	const handleClickModalContinue = () => {
+		resetAllStates();
+		setModalOpen(false);
+		setBottomSheetOpen(false);
+	};
+
+	const handleClickModalFinish = () => {
+		navigate("/");
+		setModalOpen(false);
+		setBottomSheetOpen(false);
 	};
 
 	return (
@@ -94,7 +109,16 @@ const Entry = ({ isModalOpen, setModalOpen, setBottomSheetOpen, handleSubmit, re
 			</StyledEntry>
 
 			{isModalOpen && (
-				<Modal setModalOpen={setModalOpen} setBottomSheetOpen={setBottomSheetOpen} resetAllStates={resetAllStates} />
+				<Modal>
+					<StyledRequestModal>
+						<h4>제출 완료!</h4>
+						<p>내용 검토 후 1-4일 이내 업로드 됩니다.</p>
+						<div className="modalBtnContainer">
+							<button type="button" onClick={handleClickModalContinue}>다른 카페 등록하기</button>
+							<button type="button" onClick={handleClickModalFinish}>메인으로 돌아가기</button>
+						</div>
+					</StyledRequestModal>
+				</Modal>
 			)}
 		</>
 	);
