+++ src/shared/components/BottomSheet/react-spring-bottom-sheet.css
@@ -37,7 +37,7 @@
 [data-rsbs-overlay],
 [data-rsbs-backdrop],
 [data-rsbs-root]:after {
-  z-index: 98;
+  z-index: 89;
   -ms-scroll-chaining: none;
   overscroll-behavior: none;
   touch-action: none;
