+++ src/shared/components/Modal/modalStyle.ts
@@ -1,5 +1,17 @@
 import styled from "styled-components";
 
+export const StyledModalBackground = styled.div`
+	display: block;
+	position: fixed;
+	z-index: ${({ theme }) => theme.zIndex.modal};
+	left: 0;
+	top: 0;
+	width: 100%;
+	overflow: auto;
+	height: calc(100vh + ${({ theme }) => theme.heights.header});
+	background-color: rgba(0, 0, 0, 0.4);
+`;
+
 export const StyledModal = styled.div`
 	position: absolute;
 	top: 50%;
