+++ src/shared/components/layout/header/headerStyle.ts
@@ -6,7 +6,7 @@ export const StyledHeader = styled.header<{ mainPage: boolean }>`
 	flex-direction: column;
 	width: 100%;
 	max-width: 1080px;
-	z-index: 99;
+	z-index: ${({ theme }) => theme.zIndex.header};
 	background: rgba(252, 251, 247, 0.9);
 	backdrop-filter: blur(4px);
 
@@ -16,7 +16,7 @@ export const StyledHeader = styled.header<{ mainPage: boolean }>`
 		align-items: center;
 		width: 100%;
 		max-width: 1080px;
-		height: 76px;
+		height: ${({ theme }) => theme.heights.header};
 		padding: 0 24px;
 		/* border-bottom: ${(props) => (props.mainPage ? "" : "2px solid #000")}; */
 		position: relative;
